src/test/javascript/spec/app/entities/decouverte/decouverte.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { DecouverteService } from 'app/entities/decouverte/decouverte.service';
import { IDecouverte, Decouverte } from 'app/shared/model/decouverte.model';

describe('Service Tests', () => {
  describe('Decouverte Service', () => {
    let injector: TestBed;
    let service: DecouverteService;
    let httpMock: HttpTestingController;
    let elemDefault: IDecouverte;
    let expectedResult: IDecouverte | IDecouverte[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(DecouverteService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Decouverte(0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should return a list of Decouverte', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
